admin_backend/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-login==0.5.0
pymysql==1.0.2
python-dotenv==0.19.0
werkzeug==2.0.1
pandas==1.3.33
mysqlclient==2.0.3
sqlalchemy==1.4.23
gunicorn==20.1.0
flask-migrate==3.1.0
flask-wtf==0.15.1
python-dateutil==2.8.2
requests==2.26.0
